Fix method, rename entity files
